// src/generated/api/member/models/notificationHistoryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationHistoryItemResponse } from "./notificationHistoryItemResponse";

export interface NotificationHistoryResponse {
  notifications?: NotificationHistoryItemResponse[];
  page?: number;
  pageSize?: number;
  totalCount?: number;
}